+++ libs/codegen/configs/avatarConfig.ts
@@ -1,5 +1,5 @@
-export const avatarSizes = ['m', 'l', 'xl', 'xxl', 'xxxl'];
+export const avatarSizes = ['s', 'm', 'l', 'xl', 'xxl', 'xxxl'];
 
 export const avatarPixelSizes = [16, 24, 32, 40, 48, 56];
 
-export const avatarDensePixelSizes = [16, 24, 36, 44, 48, 56];
+export const avatarDensePixelSizes = [16, 20, 24, 36, 44, 48];