+++ packages/common/internal/data/avatars.ts
@@ -3,7 +3,7 @@
  * Generated from scripts/codegen/main.ts
  */
 
-export const avatarSizes = ['m', 'l', 'xl', 'xxl', 'xxxl'] as const;
+export const avatarSizes = ['s', 'm', 'l', 'xl', 'xxl', 'xxxl'] as const;
 
 export const avatars = [
   'https://images.unsplash.com/profile-1611475141936-383e23c6cc6dimage?dpr=2&auto=format&fit=crop&w=32&h=32&q=60&crop=faces&bg=fff',