+++ packages/common/media/useAvatarSize.test.ts
@@ -1,17 +1,27 @@
 import { renderHook } from '@testing-library/react-hooks';
 
-import { getDenseAvatarPixelSize, getNormalAvatarPixelSize, useAvatarSize } from './useAvatarSize';
+import { AvatarSize } from '../types/AvatarSize';
+
+import {
+  denseScaleMap,
+  getDenseAvatarPixelSize,
+  getNormalAvatarPixelSize,
+  normalScaleMap,
+  useAvatarSize,
+} from './useAvatarSize';
 
 describe('useAvatarSize', () => {
-  it('returns normal avatar size', () => {
-    expect(getNormalAvatarPixelSize('m')).toBe(24);
+  it('returns normal avatar sizes for all sizes', () => {
+    Object.entries(normalScaleMap).forEach(([size, expectedSize]) => {
+      expect(getNormalAvatarPixelSize(size as AvatarSize)).toBe(expectedSize);
+    });
   });
 
-  it('returns dense avatar size', () => {
-    expect(getDenseAvatarPixelSize('xl')).toBe(36);
-    expect(getDenseAvatarPixelSize('xxl')).toBe(44);
+  it('returns dense avatar sizes for all sizes', () => {
+    Object.entries(denseScaleMap).forEach(([size, expectedSize]) => {
+      expect(getDenseAvatarPixelSize(size as AvatarSize)).toBe(expectedSize);
+    });
   });
-
   it('returns avatar size', () => {
     const { result } = renderHook(() => {
       return useAvatarSize('l');