+++ packages/common/media/useAvatarSize.ts
@@ -4,8 +4,15 @@ import type { AvatarDensePixelSize, AvatarPixelSize, AvatarSize } from '../types
 type AvatarDenseScaleMap = Record<AvatarSize, AvatarDensePixelSize>;
 type AvatarScaleMap = Record<AvatarSize, AvatarPixelSize>;
 
-export const denseScaleMap: AvatarDenseScaleMap = { xxxl: 48, xxl: 44, xl: 36, l: 24, m: 16 };
-export const normalScaleMap: AvatarScaleMap = { xxxl: 56, xxl: 48, xl: 40, l: 32, m: 24 };
+export const denseScaleMap: AvatarDenseScaleMap = {
+  xxxl: 48,
+  xxl: 44,
+  xl: 36,
+  l: 24,
+  m: 20,
+  s: 16,
+};
+export const normalScaleMap: AvatarScaleMap = { xxxl: 56, xxl: 48, xl: 40, l: 32, m: 24, s: 16 };
 
 export function getNormalAvatarPixelSize(size: AvatarSize) {
   return normalScaleMap[size];