+++ packages/mobile/CHANGELOG.md
@@ -8,6 +8,16 @@ All notable changes to this project will be documented in this file!
 
 <!-- template-start -->
 
+## Unreleased
+
+#### 📘 Misc
+
+- [trivial] internal: added story for visregging Frontier changes.
+
+#### 📘 Misc
+
+- [trivial] internal: added story for visregging Frontier changes.
+
 ## 5.31.4 (1/31/2024 PST)
 
 #### 🐞 Fixes